fix:  error when filename contains non-word chars.
I don't know why, but if any characters in the filename didn't match
the regex `\w`, then it would reject the filename and fail to upload.
I'm not sure why such a tight restriction was setup. I modified the
function that supposed to generate the temporary filename to replace
any undesirable characters with "-". Also added numbers, periods, and
dashes to the list of acceptable characters.